Implement a virtio driver for a TPM on a PCI bus
Note: Activation of the virtio TPM is gated by the availability of a TPM2
ACPI table providing the area for logging TPM measurements etc.

Implement a virtio driver that walks the devices on the PCI bus to detect
a matching device and then initializes it.

Since the TPM Virtio device expects the send and receive buffers at the
same time, introduce a new driver interface function 'transfer()' to use
for the Virtio driver and provide the send and receive buffes at the same
time.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>